+++ scripts/build.sh
@@ -8,18 +8,6 @@ BIN=$PWD/node_modules/.bin
 # Travis stuff
 if [ $(env | grep TRAVIS_JOB_ID ) ] ; then
 
-# coarse Node version check
- if [[ `node --version` < '5.4.1' ]]; then
-  echo "**************************************************************"
-  echo "* Your build have ben stopped because of a unsupported Node  *"
-  echo "* version. Currently Node 5 or newer is required.            *"
-  echo "*                                                            *"
-  echo "* Please fix the issue and then rerun this command.          *"
-  echo "**************************************************************"
-
-  exit 1
- fi
-
   # Install version of npm that we are locked against
   npm i -g npm@3
   # Disable the spinner, it looks bad on Travis