+++ scripts/test.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+# run in subshell to avoid polluting cwd
+(
 # if the tests fail, abort (errexit)
 set -e
 
@@ -8,6 +10,9 @@ set -e
 
 # run Mocha tests on node.js
 echo "> running tests..."
+
 ./node_modules/.bin/mocha test/specs/**/*.ts* -R spec --bail
+)
+